Move old analyses to separate module
