+++ zospy/analyses/old/mtf.py
@@ -8,7 +8,7 @@ from tempfile import mkstemp
 import pandas as pd
 
 from zospy import utils
-from zospy.analyses.base import AnalysisResult, OnComplete, new_analysis
+from zospy.analyses.old.base import AnalysisResult, OnComplete, new_analysis
 from zospy.api import constants
 from zospy.zpcore import OpticStudioSystem
