+++ zospy/analyses/old/physicaloptics.py
@@ -8,7 +8,7 @@ from typing import Literal
 import pandas as pd
 
 from zospy import utils
-from zospy.analyses.base import Analysis, AnalysisResult, OnComplete, new_analysis
+from zospy.analyses.old.base import Analysis, AnalysisResult, OnComplete, new_analysis
 from zospy.api import constants
 from zospy.zpcore import OpticStudioSystem
