+++ zospy/analyses/old/raysandspots.py
@@ -12,7 +12,7 @@ import numpy as np
 import pandas as pd
 
 import zospy.api.config as _config
-from zospy.analyses.base import AnalysisResult, AttrDict, OnComplete, new_analysis
+from zospy.analyses.old.base import AnalysisResult, AttrDict, OnComplete, new_analysis
 from zospy.api import constants
 from zospy.zpcore import OpticStudioSystem
