+++ zospy/analyses/old/surface.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import pandas as pd
 
-from zospy.analyses.base import AnalysisResult, AttrDict, OnComplete, new_analysis
+from zospy.analyses.old.base import AnalysisResult, AttrDict, OnComplete, new_analysis
 from zospy.api import constants
 from zospy.utils import zputils
 from zospy.zpcore import OpticStudioSystem
